рефакторинг кода, удалил компонент movieContainer, перенес из него логику в другие компоненты
