--- src/components/MoviesContainer/MoviesContainer.css
@@ -1,37 +1,0 @@
-.movies-container {
-    margin: auto;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    background: #FAFAFA;
-}
-
-.movies-container__button {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 90%;
-    max-width: 1140px;
-    height: 36px;
-    margin: 50px auto 80px;
-    background: #f4f4f4;
-    font-size: 12px;
-    font-style: normal;
-    font-weight: 500;
-    line-height: normal;
-    border-radius: 6px;
-    border: none;
-    cursor: pointer;
-}
-
-.movies-container__button:hover {
-    transition: 1s linear 0.03s;
-    background: rgba(244, 244, 244, 0.4);
-    color: rgba(0, 0, 0, 0.4);
-}
-
-@media (max-width: 460px) {
-    .movies-container__button {
-        margin: 50px auto 95px;
-    }
-}
