+++ src/index.css
@@ -11,7 +11,6 @@
 @import url(./components/NavTab/NavTab.css);
 @import url(./components/SearchForm/SearchForm.css);
 @import url(./components/Preloader/Preloader.css);
-@import url(./components/MoviesContainer/MoviesContainer.css);
 @import url(./components/MoviesCardList/MoviesCardList.css);
 @import url(./components/Profile/Profile.css);
 @import url(./components/FormContainer/FormContainer.css);
